busybox: rcS: create mountpoints before mount -a
mount -a doesn't create directories as necesary. While the rootfs
should contain empty directories for each mountpoint in /etc/fstab,
we're missing /dev/shm and /dev/pts if the kernel is built with

CONFIG_DEVTMPFS=y
CONFIG_DEVTMPFS_MOUNT=y

(i.e., the kernel has already mounted a devtmpfs on /dev).

I can't think of a combination of kernel .config and rootfs/fstab
contents that would break with this, but it's nice in order to
transition from a kernel without CONFIG_DEVTMPFS_MOUNT=y to one with.